pyproject.toml
[build-system]
requires = ["setuptools", "setuptools-scm"]
build-backend = "setuptools.build_meta"

[project]
name = "notion_session_database"
version = "0.1.0"
description = "A Python script to manage and update session data in Notion for neuro data analysis."
readme = "README.md"
license = { file = "LICENSE" }
authors = [{ name = "Ryan Harvey", email = "ryanh412@gmail.com" }]
keywords = ["electrophysiology", "neuroscience", "database", "notion"]
requires-python = ">=3.9"
classifiers = [
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
]

dependencies = [
    "notion-client>=2.0.0",
    "python-dotenv>=1.0.0",
    "requests>=2.25.0",
]

[project.urls]
"Homepage" = "https://github.com/ryanharvey1/notion_session_database"

[tool.setuptools.package-data]
"*" = ["*"]

[tool.setuptools.packages.find]
include = ['notion_session_database']

[tool.ruff]
line-length = 88
select = ["E", "F", "W"]

[tool.black]
target-version = ['py39']



